chore(worker): split worker fleet into fsm-sync-worker-ts and fsm-async-worker-ts (#73)
Moves apps/fsm-core-worker-ts into packages/ and splits it into two
peer packages along the sync/async dispatch-queue boundary:

- packages/fsm-sync-worker-ts (@pgfsm/sync-worker) — fsmlet, fsmscheduler,
  fsmctl; the deprecated-inprocess-approach CLI is left untouched except
  for the one import it needed fixed after the split.
- packages/fsm-async-worker-ts (@pgfsm/async-worker) — asyncOperationWorkerlet,
  asyncOperationScheduler, async-operation-ctl.

Updates every path/package reference across deno.json workspaces, CI
(npm-publish.yml gains a sync-worker/async-worker matrix), CLAUDE.md
docs, ADRs, README, and the hono-deno/example app consumers.
