+++ docs/adr/adr-003-fsm-async-operation-polyglot-actor-execution-model.md
@@ -5,7 +5,7 @@
 | Status     | Architecture: Current. Compiled-language IPC decision: **Decided, not yet implemented** (issue #55)                                        |
 | Date       | 2026-06-13 (architecture), 2026-07-22 (compiled-language IPC decision)                                                                     |
 | Deciders   | Niraj (architecture); Niraj, Claude (IPC decision)                                                                                         |
-| Affects    | `apps/fsm-core-worker-ts`, `packages/fsm-compiler-ts`, `apps/fsm-core-example`                                                             |
+| Affects    | `packages/fsm-async-worker-ts`, `packages/fsm-compiler-ts`, `apps/fsm-core-example`                                                        |
 | Supersedes | `docs/kb/kb-001-distributed-multilang-fsm.md`, `docs/specs/spec-001-compiled-lang-actor-workers.md`                                        |
 | Related    | [ADR-002](adr-002-fsm-sync-operation-worker-execution-model.md) — orchestrator tier (this ADR is its activity-tier counterpart); Issue #55 |
