+++ docs/tooling.md
@@ -35,7 +35,7 @@ tool, manifest, and lockfile apply to each part of the repo.
 
 | Ecosystem    | Tool                            | Manifest / lockfile                      | Location                                   |
 | ------------ | ------------------------------- | ---------------------------------------- | ------------------------------------------ |
-| Deno / TS    | **Deno** (JSR + npm specifiers) | `deno.json` + `deno.lock`                | repo root (workspace of 5 members)         |
+| Deno / TS    | **Deno** (JSR + npm specifiers) | `deno.json` + `deno.lock`                | repo root (workspace of 8 members)         |
 | Node         | **npm**                         | `package.json` + `package-lock.json`     | `packages/database-src`                    |
 | Rust         | **Cargo**                       | `Cargo.toml` + `Cargo.lock`              | `packages/database-src-extension/fsm_core` |
 | PG extension | **pgrx** `=0.18.1`              | pinned in `Cargo.toml`, tied to PG major | `packages/database-src-extension/fsm_core` |
@@ -138,15 +138,15 @@ All workflows live in `.github/workflows/`.
 All `uses:` actions are **pinned to commit SHAs** (with a `# vX.Y.Z` comment)
 for supply-chain safety; Renovate keeps the digests updated.
 
-| Workflow          | Trigger                                              | Tool(s)                                                                          | Purpose                                                                         |
-| ----------------- | ---------------------------------------------------- | -------------------------------------------------------------------------------- | ------------------------------------------------------------------------------- |
-| `ci.yml`          | push `main`, PR                                      | `deno fmt/lint/test` + `cargo fmt/clippy`                                        | Lint, format & unit-test gate                                                   |
-| `codeql.yml`      | push `main`, PR, weekly cron                         | `github/codeql-action`                                                           | SAST for TypeScript + Rust                                                      |
-| `reuse.yml`       | push `main`, PR                                      | `fsfe/reuse-action`                                                              | License/SPDX compliance                                                         |
-| `gitleaks.yml`    | push `main`, PR                                      | `j178/prek-action` → gitleaks                                                    | Secret scanning                                                                 |
-| `sbom.yml`        | push `main`, PR, release                             | cdxgen + grype + `upload-artifact` + `action-gh-release` + `anchore/sbom-action` | SBOM generation, CVE scan, attach SBOM to releases, dependency-graph submission |
-| `cla.yml`         | issue comment, `pull_request_target`                 | `contributor-assistant/github-action`                                            | Contributor License Agreement checks                                            |
-| `npm-publish.yml` | tag push (`compiler-v*`/`db-v*`/`worker-v*`), manual | `setup-deno` + `setup-node` + `deno pack` + `npm publish`                        | Publish TS packages to npm                                                      |
+| Workflow          | Trigger                                                                     | Tool(s)                                                                          | Purpose                                                                         |
+| ----------------- | --------------------------------------------------------------------------- | -------------------------------------------------------------------------------- | ------------------------------------------------------------------------------- |
+| `ci.yml`          | push `main`, PR                                                             | `deno fmt/lint/test` + `cargo fmt/clippy`                                        | Lint, format & unit-test gate                                                   |
+| `codeql.yml`      | push `main`, PR, weekly cron                                                | `github/codeql-action`                                                           | SAST for TypeScript + Rust                                                      |
+| `reuse.yml`       | push `main`, PR                                                             | `fsfe/reuse-action`                                                              | License/SPDX compliance                                                         |
+| `gitleaks.yml`    | push `main`, PR                                                             | `j178/prek-action` → gitleaks                                                    | Secret scanning                                                                 |
+| `sbom.yml`        | push `main`, PR, release                                                    | cdxgen + grype + `upload-artifact` + `action-gh-release` + `anchore/sbom-action` | SBOM generation, CVE scan, attach SBOM to releases, dependency-graph submission |
+| `cla.yml`         | issue comment, `pull_request_target`                                        | `contributor-assistant/github-action`                                            | Contributor License Agreement checks                                            |
+| `npm-publish.yml` | tag push (`compiler-v*`/`db-v*`/`sync-worker-v*`/`async-worker-v*`), manual | `setup-deno` + `setup-node` + `deno pack` + `npm publish`                        | Publish TS packages to npm                                                      |
 
 > **CI test scope:** `ci.yml` runs only the DB-free unit suites
 > (`packages/fsm-compiler-ts`, `apps/fsm-core-ts-hono-deno/stoker-src`). The
@@ -166,9 +166,9 @@ them to npm, and attach the CycloneDX SBOM to GitHub Releases.
 | **`npm publish`**                 | Publishes the packed tarball to the npm registry  | `npm-publish.yml` |
 | **`softprops/action-gh-release`** | Attaches the CycloneDX SBOM to GitHub Releases    | `sbom.yml`        |
 
-> Releases are tag-driven (`compiler-v*`, `db-v*`, `worker-v*`). Package
-> versions for `packages/database-src` are computed by `tsx` scripts
-> (`get-next-pkg-version.ts`).
+> Releases are tag-driven (`compiler-v*`, `db-v*`, `sync-worker-v*`,
+> `async-worker-v*`). Package versions for `packages/database-src` are computed
+> by `tsx` scripts (`get-next-pkg-version.ts`).
 
 ---
