+++ packages/fsm-logging-ts/CLAUDE.md
@@ -20,8 +20,9 @@ LogTape's `configure()` twice throws.
 - **Apps/CLIs configure**: each composition root (API `deno.ts`/`logger.ts`,
   each worker CLI, the compiler CLI) resolves levels from its own validated env
   and calls `configureLogging()` once. See
-  `apps/fsm-core-ts-hono-deno/CLAUDE.md` and `apps/fsm-core-worker-ts/CLAUDE.md`
-  for the entry points that do this.
+  `apps/fsm-core-ts-hono-deno/CLAUDE.md`,
+  `packages/fsm-sync-worker-ts/CLAUDE.md`, and
+  `packages/fsm-async-worker-ts/CLAUDE.md` for the entry points that do this.
 - **Libraries never configure**: they only `getLogger([CATEGORY.x, ...])` — e.g.
   `packages/fsm-core-db-ts/` only imports `CATEGORY`, never `configureLogging`.
 - **Env is read at the composition root**, not here — this package takes
