+++ packages/fsm-sync-worker-ts/deno.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pgfsm/worker",
+  "name": "@pgfsm/sync-worker",
   "version": "0.1.0",
   "license": "MIT",
   "exports": {
@@ -16,11 +16,8 @@
   "tasks": {
     "cli": "deno run --allow-all src/cli/fsmctl.ts",
     "dev": "deno run --allow-all --watch src/cli/fsmctl.ts",
-    "async-operation-ctl": "deno run --allow-all src/cli/async-operation-ctl.ts",
     "fsmlet": "deno run --allow-all src/cli/fsmlet.ts",
-    "async-operation-workerlet": "deno run --allow-all src/cli/async-operation-workerlet.ts",
     "fsmscheduler": "deno run --allow-all src/cli/fsmscheduler.ts",
-    "async-operation-scheduler": "deno run --allow-all src/cli/async-operation-scheduler.ts",
     "check": "deno check src/index.ts"
   }
 }
