+++ packages/fsm-sync-worker-ts/src/cli/fsmctl.ts
