+++ packages/fsm-sync-worker-ts/src/deprecated-inprocess-approach/create-and-start-fsm-worker.ts
