+++ packages/fsm-sync-worker-ts/src/deprecated-inprocess-approach/create-and-start-promise-worker.py
