+++ packages/fsm-sync-worker-ts/src/deprecated-inprocess-approach/create-and-start-promise-worker.ts
@@ -3,7 +3,7 @@ import type { DBDeps } from "@pgfsm/db";
 
 const logger = getLogger(["@pgfsm/worker", "promise"]);
 import { createPgmqQueue } from "@pgfsm/db";
-import { startFSMPromiseWorker } from "../asyncOperationWorkerlet/fsmpromiseworker.ts";
+import { startFSMPromiseWorker } from "@pgfsm/async-worker";
 import type { ActorPluginValidationResult } from "@pgfsm/compiler";
 
 export async function createAndStartPromiseWorker(
