+++ packages/fsm-sync-worker-ts/src/fsmlet/type.ts
